+++ frontend/src/components/Flow/AutoLayout.ts
@@ -1,5 +1,5 @@
-// import initialEdges from "@/_components/Flow/FlowElements/edges";
-// import initialNodes from "@/_components/Flow/FlowElements/nodes";
+ import initialEdges from "@/_components/Flow/FlowElements/edges";
+ import initialNodes from "@/_components/Flow/FlowElements/nodes";
 // import Dagre from "@dagrejs/dagre";
 // import { useCallback, useEffect } from "react";
 // import {
